Move python track store implementation to plugin layer
* simplify track repository test cases
* add unit tests for python track store implementation
