+++ OTAnalytics/plugin_intersect/simple/cut_tracks_with_sections.py
@@ -21,14 +21,13 @@ from OTAnalytics.domain.observer import OBSERVER, Subject
 from OTAnalytics.domain.section import Section, SectionId, SectionType
 from OTAnalytics.domain.track import (
     Detection,
-    PythonDetection,
-    PythonTrack,
     Track,
     TrackBuilder,
     TrackBuilderError,
     TrackClassificationCalculator,
     TrackId,
 )
+from OTAnalytics.plugin_datastore.python_track_store import PythonDetection, PythonTrack
 from OTAnalytics.plugin_intersect.shapely.mapping import ShapelyMapper
